# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Fitness Tracking Platform API
  version: 1.0.0
  description: Authoritative REST API for multi-user fitness management, daily tracking, workout/diet plan builders, progress analysis, reminders, and administration.
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server
paths:
...
  /me/cardio:
    post:
      summary: Log cardio session
      tags: [Cardio]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cardioActivityId, durationMinutes]
              properties:
                cardioActivityId: { type: integer }
                cardioDate: { type: string, format: date }
                durationMinutes: { type: integer, minimum: 1, maximum: 1440 }
                distanceKm: { type: number }
                caloriesBurned: { type: integer }
                averageHeartRate: { type: integer }
                speedKmh: { type: number }
                inclinePct: { type: number }
                notes: { type: string }
                clientOperationId: { type: string }
      responses:
        '201':
          description: Cardio logged
  /me/cardio/activities:
    get:
      summary: List available cardio activities
      tags: [Cardio]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List of cardio activities
  /me/cardio/history:
    get:
      summary: Get cardio session history
      tags: [Cardio]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200':
          description: Cardio history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items: { type: object }
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
  /admin/cardio/activities:
    get:
      summary: List all cardio activities for administration and target assignment
      tags: [Admin, Cardio]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200':
          description: List of cardio activities
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
...